Refactor/account info

+++ test/fuzz/base/UniswapV3/_UniswapV3.fuzz.t.sol
@@ -37,7 +37,7 @@ abstract contract UniswapV3_Fuzz_Test is Fuzz_Test, UniswapV3Fixture, UniswapV3A
     uint24 internal constant POOL_FEE = 100;
 
     uint256 internal constant MAX_TOLERANCE = 0.02 * 1e18;
-    uint256 internal constant MAX_FEE = 0.01 * 1e18;
+    uint64 internal constant MAX_FEE = 0.01 * 1e18;
     uint256 internal constant MIN_LIQUIDITY_RATIO = 0.99 * 1e18;
 
     /*////////////////////////////////////////////////////////////////
